fixed problems with the rule of 3/5 (copy constructor, desctructor, assignment)

+++ libj++/src/java/lang/System.cc
@@ -12,7 +12,7 @@
  	namespace lang {
  		const Class System::klass("java.lang.System", Object::klass);
 
- 		long long System::currentTimeMillis() {
+ 		jlong System::currentTimeMillis() {
  			return std::chrono::duration_cast<std::chrono::milliseconds>(std::chrono::high_resolution_clock::now().time_since_epoch()).count();
  		}
  	}
